Use azure-sql-edge instead of SqlServer
Microsoft SQL Server is only compatible with x86-64 host architecture.
This means we cannot run Docker Compose in ARM environment, most notably
in macOS on Apple Silicon. While there is an emulation mode for x86-64
images in macOS (Rosetta 2), it is not good enough - SQL Server x86-64
image crashed once we start using it.

Alternative is to use Azure SQL Edge
https://azure.microsoft.com/en-gb/products/azure-sql/edge/#product-overview

It may have some limitations in terms of supported features compared
to the full SQL Server, but it is perfectly adequate for the purpose
of this demo application.